Extracted default tick mark formatter to separate module + added tests
